Create OpenGL Display

+++ Source/BaseObject.cpp
@@ -10,7 +10,7 @@
 
 #include "BaseObject.h"
 
-BodyPoint::BodyPoint(double x, double y, double z)
+BodyPoint::BodyPoint(float x, float y, float z)
 {
     bodyPoint = Vector(x, y, z);
 }
